+++ addons/website_slides/models/__init__.py
@@ -5,6 +5,7 @@ from . import ir_http
 from . import gamification_challenge
 from . import slide_slide
 from . import slide_question
+from . import slide_embed
 from . import slide_channel
 from . import slide_channel_tag
 from . import slide_slide_resource
